Move main plugin class to a PSR-4 file
